# documentation/docs/content/openAPI/spring_backend-openapi.yaml
openapi: "3.0.3"
info:
  title: "spring_backend API"
  description: "spring_backend API"
  version: "1.0.0"
servers:
  - url: "https://spring_backend"
paths:
...
  /api/django/ml/{modelId}:
    get:
      summary: "GET api/django/ml/{modelId}"
      operationId: "getValidSolutions"
      parameters:
        - name: "modelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/MLConfigurationDTO"
  /api/django/tsd/{modelId}:
    get:
      summary: "GET api/django/tsd/{modelId}"
      operationId: "getValidTSDConfiguration"
      parameters:
        - name: "modelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/TSDConfigurationDTO"
  /api/django/tsd/{modelId}/tsd_config/{configId}/trainData:
    get:
      summary: "GET api/django/tsd/{modelId}/tsd_config/{configId}/trainData"
      operationId: "getTrainDataOfConfiguration"
      parameters:
        - name: "modelId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "configId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/TrainDataDTO"
  /api/logs/{trackId}:
    get:
      summary: "GET api/logs/{trackId}"
      operationId: "getLogsForTrack"
      parameters:
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "page_size"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
        - name: "page_number"
          in: "query"
          required: false
          schema:
            type: "integer"
            format: "int32"
        - name: "search_query"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/LogMessageDTO"
  /api/logs/{trackId}/log_size:
    get:
      summary: "GET api/logs/{trackId}/log_size"
      operationId: "getLogSize"
      parameters:
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "search_query"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "integer"
                format: "int32"
  /api/logs/{trackId}/logSessions:
    get:
      summary: "GET api/logs/{trackId}/logSessions"
      operationId: "getLogSessions"
      parameters:
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/LogSessionDTO"
  /api/logs/{trackId}/logSession:
    get:
      summary: "GET api/logs/{trackId}/logSession"
      operationId: "getAllByLogSession"
      parameters:
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "session"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/LogMessageDTO"
  /api/logs/{trackId}/logGraph:
    get:
      summary: "GET api/logs/{trackId}/logGraph"
      operationId: "getLogGraph"
      parameters:
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "session"
          in: "query"
          required: false
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
...
components:
  schemas:
...
    Instant:
      type: "object"
      properties: { }
    MLModelsDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        created_at:
          $ref: "#/components/schemas/Instant"
    TSDModelDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        created_at:
          $ref: "#/components/schemas/Instant"
    MLConfigurationDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        mlmodel:
          $ref: "#/components/schemas/MLModelsDTO"
        created_at:
          $ref: "#/components/schemas/Instant"
        training_time:
          type: "integer"
          format: "int32"
        training_iterations:
          type: "integer"
          format: "int32"
        accuracy:
          type: "number"
          format: "float"
        max_length:
          type: "integer"
          format: "int32"
    TSDConfigurationDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        mlmodel_id:
          $ref: "#/components/schemas/TSDModelDTO"
        created_at:
          $ref: "#/components/schemas/Instant"
        max_length:
          type: "integer"
          format: "int32"
        levelConfig:
          type: "string"
    TrainDataDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
    LogMessageDTO:
      type: "object"
      properties:
        message:
          type: "string"
        time:
          type: "string"
        sendSession:
          type: "string"
        dataSetName:
          type: "string"
    LogSessionDTO:
      type: "object"
      properties:
        session:
          type: "string"
    JsonNode:
      type: "object"
      properties: { }
...